Move the day picker customization into a wrapper component because it's growing
